+++ hc/front/tests/test_ping_details.py
@@ -268,6 +268,16 @@ class PingDetailsTestCase(BaseTestCase):
         r = self.client.get(self.url)
         self.assertContains(r, "please check back later", status_code=200)
 
+    @override_settings(S3_BUCKET="test-bucket")
+    @patch("hc.api.models.get_object")
+    def test_it_handles_missing_object_email(self, get_object: Mock) -> None:
+        Ping.objects.create(owner=self.check, n=1, scheme="email", object_size=1000)
+        get_object.return_value = None
+
+        self.client.login(username="alice@example.org", password="password")
+        r = self.client.get(self.url)
+        self.assertContains(r, "please check back later", status_code=200)
+
     @override_settings(S3_BUCKET=None)
     def test_it_handles_missing_s3_credentials(self) -> None:
         Ping.objects.create(owner=self.check, n=1, object_size=1000)